Fixed Issue with Navigation!
Assisted with fixing an navigation issue!
