+++ src/sites/Music.js
@@ -1,6 +1,3 @@
-
-
-
 export default function Music(){
     return(
         <div>
